move all MTU structs and constants to mtu.rs
