+++ quic/s2n-quic-core/src/path/mod.rs
@@ -5,11 +5,7 @@ use crate::{
     event,
     inet::{IpV4Address, IpV6Address, SocketAddress, SocketAddressV4, SocketAddressV6},
 };
-use core::{
-    fmt,
-    fmt::{Display, Formatter},
-    num::NonZeroU16,
-};
+use core::fmt;
 
 #[cfg(any(test, feature = "generator"))]
 use bolero_generator::*;
@@ -18,42 +14,7 @@ pub mod ecn;
 pub mod migration;
 pub mod mtu;
 
-//= https://www.rfc-editor.org/rfc/rfc9000#section-14
-//# QUIC MUST NOT be used if the network path cannot support a
-//# maximum datagram size of at least 1200 bytes.
-// TODO: rename to MINIMUM_DATAGRAM_SIZE
-pub const MINIMUM_MTU: u16 = 1200;
-
-// TODO decide on better defaults
-// Safety: 1500 is greater than zero
-pub const DEFAULT_MAX_MTU: MaxMtu = MaxMtu(unsafe { NonZeroU16::new_unchecked(1500) });
-const DEFAULT_MIN_MTU: MinMtu = MinMtu(unsafe { NonZeroU16::new_unchecked(MIN_ALLOWED_MAX_MTU) });
-const DEFAULT_INITIAL_MTU: InitialMtu =
-    InitialMtu(unsafe { NonZeroU16::new_unchecked(MIN_ALLOWED_MAX_MTU) });
-
-// Length is the length in octets of this user datagram  including  this
-// header and the data. (This means the minimum value of the length is
-// eight.)
-// See https://www.rfc-editor.org/rfc/rfc768.txt
-pub const UDP_HEADER_LEN: u16 = 8;
-
-// IPv4 header ranges from 20-60 bytes, depending on Options
-pub const IPV4_MIN_HEADER_LEN: u16 = 20;
-// IPv6 header is always 40 bytes, plus extensions
-pub const IPV6_MIN_HEADER_LEN: u16 = 40;
-
-// The minimum allowed Max MTU is the minimum UDP datagram size of 1200 bytes plus
-// the UDP header length and minimal IP header length
-const fn const_min(a: u16, b: u16) -> u16 {
-    if a < b {
-        a
-    } else {
-        b
-    }
-}
-// TODO: rename to MINIMUM_MTU
-const MIN_ALLOWED_MAX_MTU: u16 =
-    MINIMUM_MTU + UDP_HEADER_LEN + const_min(IPV4_MIN_HEADER_LEN, IPV6_MIN_HEADER_LEN);
+pub use mtu::*;
 
 // Initial PTO backoff multiplier is 1 indicating no additional increase to the backoff.
 pub const INITIAL_PTO_BACKOFF: u32 = 1;
@@ -289,80 +250,6 @@ impl Handle for Tuple {
         }
     }
 }
-
-macro_rules! impl_mtu {
-    ($name:ident, $default:expr) => {
-        #[derive(Clone, Copy, Debug, PartialEq)]
-        pub struct $name(NonZeroU16);
-
-        impl $name {
-            /// The minimum value required for path MTU
-            pub const MIN: Self = Self(unsafe { NonZeroU16::new_unchecked(MIN_ALLOWED_MAX_MTU) });
-
-            /// The Packetization Layer Path MTU, the largest size of a QUIC datagram that can be
-            /// sent on a path. This does not include the size of UDP and IP headers.
-            pub fn plpmtu(&self, ip_header_len: u16) -> u16 {
-                u16::from(*self) - UDP_HEADER_LEN - ip_header_len
-            }
-        }
-
-        impl Default for $name {
-            fn default() -> Self {
-                $default
-            }
-        }
-
-        impl TryFrom<u16> for $name {
-            type Error = MtuError;
-
-            fn try_from(value: u16) -> Result<Self, Self::Error> {
-                if value < MIN_ALLOWED_MAX_MTU {
-                    return Err(MtuError(MIN_ALLOWED_MAX_MTU.try_into().unwrap()));
-                }
-
-                Ok($name(value.try_into().expect(
-                    "Value must be greater than zero according to the check above",
-                )))
-            }
-        }
-
-        impl From<$name> for usize {
-            #[inline]
-            fn from(value: $name) -> Self {
-                value.0.get() as usize
-            }
-        }
-
-        impl From<$name> for u16 {
-            #[inline]
-            fn from(value: $name) -> Self {
-                value.0.get()
-            }
-        }
-    };
-}
-
-impl_mtu!(MaxMtu, DEFAULT_MAX_MTU);
-impl_mtu!(InitialMtu, DEFAULT_INITIAL_MTU);
-impl_mtu!(MinMtu, DEFAULT_MIN_MTU);
-
-#[derive(Debug)]
-pub struct MtuError(NonZeroU16);
-
-impl Display for MtuError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "Mtu must be at least {}", self.0)
-    }
-}
-
-// impl core::cmp::PartialOrd<InitialMtu> for MaxMtu {
-//     fn partial_cmp(&self, other: &InitialMtu) -> Option<Ordering> {
-//         self.0.partial_cmp(&other.0)
-//     }
-// }
-
-#[cfg(feature = "std")]
-impl std::error::Error for MtuError {}
 
 //= https://www.rfc-editor.org/rfc/rfc9308#section-8.1
 //# Some UDP protocols are vulnerable to reflection attacks, where an
