+++ quic/s2n-quic-core/src/path/mtu/tests.rs
@@ -8,7 +8,6 @@ use crate::{
     frame::Frame,
     inet::{IpV4Address, SocketAddressV4},
     packet::number::PacketNumberSpace,
-    path::MIN_ALLOWED_MAX_MTU,
     recovery::congestion_controller::testing::mock::CongestionController,
     time::{clock::testing::now, timer::Provider as _},
     transmission::{
